Convert `#include <xstring>` to `#include <string>` in nearby files.
`<xstring>` is the MSVC internal implementation file for string. This makes it contingent on the MSVC STL.

`<string>` provides the same declarations, but does not expose the MSVC-specific implementation details. This implementation is guaranteed to meet the C++ standard exactly.

This change updates includes of `<xstring>` to be includes of `<string>`.

PiperOrigin-RevId: 483696244